vendor: Switch the search bar round on click
We should thank google every day of our life for doing stuff in a horrible way. This may be a really ugly method but this saves us from a unneccesary fork.
Let's use the same value of the search bar overlay to keep it consistent. And also modify the search_bar_margin to use the same values as Settings

Change-Id: I3940a72d55a93eed78725876dbf5efc6eca2fc4f
Signed-off-by: spezi77 <spezi7713@gmx.net>